+++ pkg/scalers/kafka_scaler.go
@@ -43,10 +43,11 @@ type kafkaMetadata struct {
 	password string
 
 	// TLS
-	enableTLS bool
-	cert      string
-	key       string
-	ca        string
+	enableTLS   bool
+	cert        string
+	key         string
+	keyPassword string
+	ca          string
 
 	scalerIndex int
 }
@@ -141,6 +142,11 @@ func parseKafkaAuthParams(config *ScalerConfig, meta *kafkaMetadata) error {
 			meta.ca = config.AuthParams["ca"]
 			meta.cert = config.AuthParams["cert"]
 			meta.key = config.AuthParams["key"]
+			if value, found := config.AuthParams["keyPassword"]; found {
+				meta.keyPassword = value
+			} else {
+				meta.keyPassword = ""
+			}
 			meta.enableTLS = true
 		} else if val != "disable" {
 			return fmt.Errorf("err incorrect value for TLS given: %s", val)
@@ -280,7 +286,7 @@ func getKafkaClients(metadata kafkaMetadata) (sarama.Client, sarama.ClusterAdmin
 
 	if metadata.enableTLS {
 		config.Net.TLS.Enable = true
-		tlsConfig, err := kedautil.NewTLSConfig(metadata.cert, metadata.key, metadata.ca)
+		tlsConfig, err := kedautil.NewTLSConfigWithPassword(metadata.cert, metadata.key, metadata.keyPassword, metadata.ca)
 		if err != nil {
 			return nil, nil, err
 		}
@@ -326,6 +332,12 @@ func (s *kafkaScaler) getTopicPartitions() (map[string][]int32, error) {
 		if err != nil {
 			return nil, fmt.Errorf("error listing cg offset: %s", err)
 		}
+
+		if listCGOffsetResponse.Err > 0 {
+			errMsg := fmt.Errorf("error listing cg offset: %s", listCGOffsetResponse.Err.Error())
+			kafkaLog.Error(errMsg, "")
+		}
+
 		for topicName := range listCGOffsetResponse.Blocks {
 			topicsToDescribe = append(topicsToDescribe, topicName)
 		}
@@ -344,6 +356,10 @@ func (s *kafkaScaler) getTopicPartitions() (map[string][]int32, error) {
 
 	topicPartitions := make(map[string][]int32, len(topicsMetadata))
 	for _, topicMetadata := range topicsMetadata {
+		if topicMetadata.Err > 0 {
+			errMsg := fmt.Errorf("error describing topics: %s", topicMetadata.Err.Error())
+			kafkaLog.Error(errMsg, "")
+		}
 		partitionMetadata := topicMetadata.Partitions
 		partitions := make([]int32, len(partitionMetadata))
 		for i, p := range partitionMetadata {
@@ -359,6 +375,10 @@ func (s *kafkaScaler) getConsumerOffsets(topicPartitions map[string][]int32) (*s
 	if err != nil {
 		return nil, fmt.Errorf("error listing consumer group offsets: %s", err)
 	}
+	if offsets.Err > 0 {
+		errMsg := fmt.Errorf("error listing consumer group offsets: %s", offsets.Err.Error())
+		kafkaLog.Error(errMsg, "")
+	}
 	return offsets, nil
 }
 
@@ -369,6 +389,11 @@ func (s *kafkaScaler) getLagForPartition(topic string, partitionID int32, offset
 		kafkaLog.Error(errMsg, "")
 		return 0, errMsg
 	}
+	if block.Err > 0 {
+		errMsg := fmt.Errorf("error finding offset block for topic %s and partition %d: %s", topic, partitionID, offsets.Err.Error())
+		kafkaLog.Error(errMsg, "")
+	}
+
 	consumerOffset := block.Offset
 	if consumerOffset == invalidOffset && s.metadata.offsetResetPolicy == latest {
 		retVal := int64(1)